e9. started setting up engine. Rather than an object oriented approach to entity-component organziation, this will use a more id driven method following a blog post from auto desk about their stingray engine. The ids in that method involve a generation part and an id part. The generation part is used as a differenciation tool between old references to the same item in lists of entities and types of components (scripts, transforms, rigid bodies, .etc). modified engine project settings, removed default files, setup primitive types to simpler names, created id system and some useful macros. Defined some initial methods needed to implement entities and components.
Diverged from the videos in convention here. Types are caps if non primitive (so u32 is an unsigned 32 bit int, but IdType is a non trivial definition, so it gets capitalized). I also prefer camel casing to snake case, so I'll use that instead, however, I do follow the snake case convention on macros.

I'll likely come back to this before e10 and swap out the function declarations from the old way (ret_type name def) for the new way in c++20 (auto name -> ret_type def) as I kind of like that.