e19. Added functionality to build user game code from editor via visual studio. Added a button and keybinding to start build process, unload game code dll via c++ engine, make user visual studio instance build the game code solution, and load the produced dll for use in the editor.
Deviated from tutorial videos in one spot. EnvDTE.ExecuteCommand seems to not block in my case (returns and continues before callback associated with completion is called). This caused a race condition. Used a different function which has an option to block to be sure it will block. This was done in the VisualStudio.BuildSolution function.