e20.2. Added scipt component, script view, and dropdown for adding components to game objects from the editor. New scripts can be added via the new script button and immediately populate the addcomponent script dropdown so the scripts are ready to be attached to game objects.
Fixed a bug from last commit in EngineAPI.cpp where a function was missnamed (GetScriptNAmes -> GetScriptNames).
Modified editor .csproj file to deal with odd bug. If BaseOutputPath is ../ or ., new classes automatically add compilation items to the .csproj file. This then causes NETSDK1022. Changed it to $(ProjectDir)../ which seems to have avoided the bug.